21442: Users find it difficult to locate flow.js on our flow document page.
I've added a few inline links in the doc body text. I also decided to add a Resources section with the script name, description, and URL. I think this will be a useful thing to have on tutorial pages to help users find all the scripts needed for a doc.